Added no username to admin bot

+++ src/plugins/xatta-realtor-admin-bot/server/bot/components/inlineCommands/enterCommand.js
@@ -1,4 +1,5 @@
 const createNewAuthor = require('../../../../botUtils/userController/createNewAuthor');
+const { NO_USERNAME } = require('../../../../botUtils/errorHandlers');
 
 /**
  * @param user
@@ -14,6 +15,7 @@ const sendWelcomMessage = async ({ user, localisation }) =>
  */
 module.exports = async (msg) => {
     const { user, localisation } = msg;
+    if (!process.env.DEVELOPMENT && !msg.user.username) return NO_USERNAME(msg);
     const messageId = msg.message?.message_id || msg.message_id;
     await strapi.bots.admin.deleteMessage(user.telegramID, messageId);
